Add some shorthand setters to nodes
I'm still confused by why my FM example fails to work.  One thing I've been a little concerned about is that if a user wants simply to set the value of a node's AudioParam he must first read that parameter.  It seems to me that this should be hidden inside a simple setter function.

Accordingly I've added setters for gain (gainNode) and frequency and detune (oscillatorNode).