Add experimental unsafeConnectParam
The safe version from a previous commit was so safe that it didn't work.  My suspicion is that this was caused by PureScript, being pure, assuming that JavaScript had similar purity.  Do notation transpiles to a set of assignment statements but I suspect that connecting to an audio parameter requires the original parameter, not a copy.